Improved version of Msp430FrequencyControl component.  - Added a call to Pmm interface.  - Cleaner commands, easier to understand  - Changed the frequency representation unit to Hz, now frequency variables are in uint32_t  - Changed return types and variable types of some commands to better suit the value in question  - Added header file with constants  - Commented setDCORange, and automatically put it to RSELx=4. (I'm testing with input frequency of 3MHz to the setMCLKFreq command, this range I know that works)
Problems:
DCO ony settles with the right RSELx value. Still trying to find the way to do this.
The definition of the XT1_FREQ, which is the external crystal frequency value, is not being accepted by the compiler, and it's causing an overflow. Defining it to 32768 appears as -32768.
Must find how to solve this.

Previous testings with the frequency value in kHz, and XT1_FREQ defined to 32 worked fine.