DVS test App:
Micaz requests te start of the calcule of a pre-determined number of the fibonacci sequence.
MoteISTx5 replys in a sort of handshake, and sends the message when the asked number of the sequence are calculated.

Mica has a deadline to receive the taskdone message, if MoteISTx5 fails to send the message in time mica sends the DEADLINE_MISS msg.

To do:

- put togheter the interface to set the frequency whenever a deadline cant be met with the actual one
- make periodic reading to the ADC to measure the current being draw
- write to flash the adc readings and bunddle them with a state structure that has: the actual Vcore,
supply voltage (of the batteries, also read by the adc), state of the mote (waiting or processing)