Application to simulate a real-time system
Receives requests from a Mica mote with an number of iterations to perform and a deadline to do it
MoteIST adapts its frequency to meet the deadlines in 50% the time of the deadline (i.e. the lowest frequency that allows to)

The number of iterations changes up and down so that we can see how it behaves.

To do:
- insert the ADC measurements to get the total power of the system